+++ server/src/app.js
@@ -5,6 +5,7 @@ const morgan = require("morgan");
 const cookieParser = require("cookie-parser");
 
 const authRoutes = require("./routes/auth.routes");
+const warehouseRoutes = require("./routes/warehouse.routes");
 
 const notFound = require("./middlewares/notFound.middleware");
 const errorMiddleware = require("./middlewares/error.middleware");
@@ -31,7 +32,7 @@ app.get("/api/v1/health", (req, res) => {
 
 
 app.use("/api/v1/auth", authRoutes);
-
+app.use("/api/v1/warehouses", warehouseRoutes);
 
 app.use(notFound);
 app.use(errorMiddleware);